Move current components into components folder
